feat(realtime-server): add env corresponding to notification instance status
We didn't have a way to know on deployments how is the notification instance status. We added the extra REALTIME_SERVER_ENABLED boolean value so worker and server knows at all times when the notification instance is up so client also behaves as it should. This had to be reflected in the chart as well.